requirements.txt
streamlit
tensorflow
opencv-python-headless
Pillow
numpy
altair
pandas
streamlit-webrtc
